+++ R/content-tools.R
@@ -1,13 +1,9 @@
 # Results a content list
-invoke_tools <- function(turn, tools) {
-  if (length(tools) == 0) {
-    return()
-  }
+invoke_tools <- function(turn) {
   tool_requests <- extract_tool_requests(turn@contents)
 
-  lapply(tool_requests, function(call) {
-    fun <- tools[[call@name]]@fun
-    result <- invoke_tool(fun, call@arguments, call@id)
+  lapply(tool_requests, function(request) {
+    result <- invoke_tool(request)
 
     if (promises::is.promise(result@value)) {
       cli::cli_abort(c(
@@ -22,17 +18,12 @@ invoke_tools <- function(turn, tools) {
 
 on_load(
   invoke_tools_async <- coro::async(function(turn, tools) {
-    if (length(tools) == 0) {
-      return()
-    }
-
     tool_requests <- extract_tool_requests(turn@contents)
 
     # We call it this way instead of a more natural for + await_each() because
     # we want to run all the async tool calls in parallel
-    result_promises <- lapply(tool_requests, function(call) {
-      fun <- tools[[call@name]]@fun
-      invoke_tool_async(fun, call@arguments, call@id)
+    result_promises <- lapply(tool_requests, function(request) {
+      invoke_tool_async(request)
     })
 
     promises::promise_all(.list = result_promises)
@@ -44,35 +35,50 @@ extract_tool_requests <- function(contents) {
   contents[is_tool_request]
 }
 
+new_tool_result <- function(request, result = NULL, error = NULL) {
+  check_exclusive(result, error)
+
+  if (!is.null(error)) {
+    ContentToolResult(error = error, request = request)
+  } else if (S7_inherits(result, ContentToolResult)) {
+    set_props(result, request = request)
+  } else {
+    ContentToolResult(value = result, request = request)
+  }
+}
+
 # Also need to handle edge cases: https://platform.openai.com/docs/guides/function-calling/edge-cases
-invoke_tool <- function(fun, arguments, id) {
-  if (is.null(fun)) {
-    return(ContentToolResult(id = id, error = "Unknown tool"))
+invoke_tool <- function(request) {
+  if (is.null(request@tool)) {
+    return(new_tool_result(request, error = "Unknown tool"))
   }
 
   tryCatch(
-    ContentToolResult(id, do.call(fun, arguments)),
+    {
+      result <- do.call(request@tool@fun, request@arguments)
+      new_tool_result(request, result)
+    },
     error = function(e) {
       # TODO: We need to report this somehow; it's way too hidden from the user
-      ContentToolResult(id, error = e)
+      new_tool_result(request, error = e)
     }
   )
 }
 
 on_load(
-  invoke_tool_async <- coro::async(function(fun, arguments, id) {
-    if (is.null(fun)) {
-      return(ContentToolResult(id = id, error = "Unknown tool"))
+  invoke_tool_async <- coro::async(function(request) {
+    if (is.null(request@tool)) {
+      return(new_tool_result(request, error = "Unknown tool"))
     }
 
     tryCatch(
       {
-        result <- await(do.call(fun, arguments))
-        ContentToolResult(id, result)
+        result <- await(do.call(request@tool@fun, request@arguments))
+        new_tool_result(request, result)
       },
       error = function(e) {
         # TODO: We need to report this somehow; it's way too hidden from the user
-        ContentToolResult(id, error = e)
+        new_tool_result(request, error = e)
       }
     )
   })