+++ R/content.R
@@ -1,3 +1,4 @@
+#' @include tools-def.R
 #' @include utils-S7.R
 NULL
 
@@ -70,6 +71,10 @@ contents_markdown <- new_generic("contents_markdown", "content")
 #' * `ContentToolRequest`: a request to perform a tool call (sent by the
 #'    assistant).
 #' * `ContentToolResult`: the result of calling the tool (sent by the user).
+#'   This object is automatically created from the value returned by calling the
+#'   [tool()] function. Alternatively, expert users can return a
+#'   `ContentToolResult` from a [tool()] function to include additional data or
+#'   to customize the display of the result.
 #'
 #' @export
 #' @return S7 objects that all inherit from `Content`
@@ -175,16 +180,20 @@ method(contents_markdown, ContentImageInline) <- function(content) {
 
 #' @rdname Content
 #' @export
-#' @param id Tool call id (used to associate a request and a result)
+#' @param id Tool call id (used to associate a request and a result).
+#'   Automatically managed by \pkg{ellmer}.
 #' @param name Function name
 #' @param arguments Named list of arguments to call the function with.
+#' @param tool ellmer automatically matches a tool request to the tools defined
+#'   for the chatbot. If `NULL`, the request did not match a defined tool.
 ContentToolRequest <- new_class(
   "ContentToolRequest",
   parent = Content,
   properties = list(
     id = prop_string(),
     name = prop_string(),
-    arguments = class_list
+    arguments = class_list,
+    tool = NULL | ToolDef
   )
 )
 method(format, ContentToolRequest) <- function(x, ...) {
@@ -202,11 +211,16 @@ method(format, ContentToolRequest) <- function(x, ...) {
 #' @param error The error message, as a string, or the error condition thrown
 #'   as a result of a failure when calling the tool function. Must be `NULL`
 #'   when the tool call is successful.
+#' @param extra Optional additional data associated with the tool result that
+#'   isn't included in the `value` that's shown to the LLM. Useful for including
+#'   additional data for displaying the tool result in a client, like a Shiny
+#'   app, without including the data in the response to the LLM.
+#' @param request The [ContentToolRequest] associated with the tool result,
+#'   automatically added by \pkg{ellmer} when evaluating the tool call.
 ContentToolResult <- new_class(
   "ContentToolResult",
   parent = Content,
   properties = list(
-    id = prop_string(),
     value = class_any,
     error = new_property(
       class = NULL | class_character | new_S3_class("condition"),
@@ -223,7 +237,9 @@ ContentToolResult <- new_class(
           "."
         )
       }
-    )
+    ),
+    extra = class_list,
+    request = NULL | ContentToolRequest
   )
 )
 method(format, ContentToolResult) <- function(x, ...) {
@@ -232,7 +248,7 @@ method(format, ContentToolResult) <- function(x, ...) {
   } else {
     value <- x@value
   }
-  cli::format_inline("[{.strong tool result}  ({x@id})]: {value}")
+  cli::format_inline("[{.strong tool result}  ({x@request@id})]: {value}")
 }
 
 tool_errored <- function(x) !is.null(x@error)