+++ tests/testthat/test-content-tools.R
@@ -1,49 +1,161 @@
-test_that("invoke_tool returns a tool_result", {
-  res <- invoke_tool(function() 1, list(), id = "x")
+test_that("invoke_tool returns a ContentToolResult", {
+  tool <- tool(function() 1, "A tool", .name = "my_tool")
+
+  res <- invoke_tool(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(),
+      tool = tool
+    )
+  )
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_equal(res@error, NULL)
   expect_false(tool_errored(res))
   expect_equal(res@value, 1)
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool)
+  expect_equal(res@request@arguments, list())
 
-  res <- invoke_tool(function() 1, list(x = 1), id = "x")
+  res <- invoke_tool(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(x = 1),
+      tool = tool
+    )
+  )
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_s3_class(res@error, "condition")
   expect_true(tool_errored(res))
   expect_equal(tool_error_string(res), "unused argument (x = 1)")
   expect_equal(res@value, NULL)
+  expect_equal(res@extra, list())
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool)
+  expect_equal(res@request@arguments, list(x = 1))
 
-  res <- invoke_tool(NULL, list(x = 1), id = "x")
+  res <- invoke_tool(
+    ContentToolRequest(
+      id = "x",
+      arguments = list(x = 1),
+      name = "my_tool"
+    )
+  )
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_equal(res@error, "Unknown tool")
   expect_equal(tool_error_string(res), "Unknown tool")
   expect_true(tool_errored(res))
   expect_equal(res@value, NULL)
+  expect_equal(res@extra, list())
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, NULL)
+  expect_equal(res@request@arguments, list(x = 1))
+
+  tool_ctr <- tool(
+    function() ContentToolResult(value = 1, extra = list(a = 1)),
+    "A tool that returns ContentToolResult",
+    .name = "my_tool"
+  )
+  res <- invoke_tool(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(),
+      tool = tool_ctr
+    )
+  )
+  expect_s3_class(res, "ellmer::ContentToolResult")
+  expect_equal(res@error, NULL)
+  expect_false(tool_errored(res))
+  expect_equal(res@value, 1)
+  expect_equal(res@extra, list(a = 1))
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool_ctr)
+  expect_equal(res@request@arguments, list())
 })
 
-test_that("invoke_tool_async returns a tool_result", {
-  res <- sync(invoke_tool_async(function() 1, list(), id = "x"))
+test_that("invoke_tool_async returns a ContentToolResult", {
+  tool <- tool(function() 1, "A tool", .name = "my_tool")
+
+  res <- sync(invoke_tool_async(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(),
+      tool = tool
+    )
+  ))
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_equal(res@error, NULL)
   expect_false(tool_errored(res))
   expect_equal(res@value, 1)
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool)
+  expect_equal(res@request@arguments, list())
 
-  res <- sync(invoke_tool_async(function() 1, list(x = 1), id = "x"))
+  res <- sync(invoke_tool_async(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(x = 1),
+      tool = tool
+    )
+  ))
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_s3_class(res@error, "condition")
-  expect_equal(tool_error_string(res), "unused argument (x = 1)")
   expect_true(tool_errored(res))
+  expect_equal(tool_error_string(res), "unused argument (x = 1)")
   expect_equal(res@value, NULL)
+  expect_equal(res@extra, list())
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool)
+  expect_equal(res@request@arguments, list(x = 1))
 
-  res <- sync(invoke_tool_async(NULL, list(x = 1), id = "x"))
+  res <- sync(invoke_tool_async(
+    ContentToolRequest(
+      id = "x",
+      arguments = list(x = 1),
+      name = "my_tool"
+    )
+  ))
   expect_s3_class(res, "ellmer::ContentToolResult")
-  expect_equal(res@id, "x")
   expect_equal(res@error, "Unknown tool")
   expect_equal(tool_error_string(res), "Unknown tool")
   expect_true(tool_errored(res))
   expect_equal(res@value, NULL)
+  expect_equal(res@extra, list())
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, NULL)
+  expect_equal(res@request@arguments, list(x = 1))
+
+  tool_ctr <- tool(
+    function() ContentToolResult(value = 1, extra = list(a = 1)),
+    "A tool that returns ContentToolResult",
+    .name = "my_tool"
+  )
+  res <- sync(invoke_tool_async(
+    ContentToolRequest(
+      id = "x",
+      name = "my_tool",
+      arguments = list(),
+      tool = tool_ctr
+    )
+  ))
+  expect_s3_class(res, "ellmer::ContentToolResult")
+  expect_equal(res@error, NULL)
+  expect_false(tool_errored(res))
+  expect_equal(res@value, 1)
+  expect_equal(res@extra, list(a = 1))
+  expect_s3_class(res@request, "ellmer::ContentToolRequest")
+  expect_equal(res@request@id, "x")
+  expect_equal(res@request@tool, tool_ctr)
+  expect_equal(res@request@arguments, list())
 })